LowRankIntegrators

Documentation for LowRankIntegrators.

LowRankIntegrators.DLRIntegrator

LowRankIntegrators.DLRSolution

LowRankIntegrators.MatrixDEProblem

LowRankIntegrators.MatrixDataProblem

LowRankIntegrators.MatrixHybridProblem

LowRankIntegrators.solve

LowRankIntegrators.DLRIntegrator — Type

Integrator computing solution to a dynamic low rank approximation problem

LowRankIntegrators.DLRSolution — Type

Solution object that tracks the evolution of a low rank approximation

LowRankIntegrators.MatrixDEProblem — Type

Problem of approximating the solution of a matrix differential equation
dy/dt = f(t,y) with y(0) = y₀ on [t0,tf] with a low rank factorization 
y(t) ≈ u(t) = U(t)S(t)V(t)' where U(t) and V(t) are orthonormal bases approximating 
range and co-range of y(t).

LowRankIntegrators.MatrixDataProblem — Type

Problem of tracking the low rank decomposition u(t) = U(t)S(t) V(t)' (or U(t)Z(t)') of a 
time-dependent (or streamed) matrix y(t) with t ∈ [t_0, t_f].

LowRankIntegrators.MatrixHybridProblem — Type

Problem of identifying optimal projection basis U(t) such that we get good reconstruction
y(t) = U(t) Z(t)' where dZ/dt = f(Z,U,t).

LowRankIntegrators.solve — Method

solves the given problem with the specified algorithm and step size